reorganized to sample.py/sample_test.py to match tfp.mcmc
